refactor: remove redundant Knowledge re-export
Address review feedback — the module was both importing Knowledge from
@/types/knowledge and re-exporting it from the same path. Drop the
re-export and point the two consumers (KnowledgeBaseSection,
UpdateArtistInfoSuccess) at the canonical @/types/knowledge source.
